refactor car block styles into own file
