+++ src/components/screens/carScroller/carScrollerDesktop.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 
 import { CarType } from '../../../types/cars'
-import { CarBlock } from '../../organisms/carBlock'
+import { CarBlock } from '../../organisms/carBlock/carBlock'
 import { DesktopScroller } from '../../organisms/desktopScroller'
 
 type Props = {
